Fix Dog dao

+++ src/main/java/com/k9club/api/dao/DogDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DogDao extends JpaRepository<Dog, Integer> {
+public interface DogDao extends JpaRepository<Dog, Long> {
 }
